refactor: 벽 설치를 별도 메서드로 분시
